+++ src/components/custom-field-input-multiple-choice/custom-field-input-multiple-choice.test.jsx
@@ -1,13 +1,22 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import {
+  fireEvent,
   render,
   screen,
 } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import CustomFieldInputMultipleChoice from './custom-field-input-multiple-choice.jsx';
 
 describe('<CustomFieldInputMultipleChoice>', () => {
   const requiredProps = {
+    choices: [{
+      id: '1',
+      label: 'Choice 1',
+    }, {
+      id: '2',
+      label: 'Choice 2',
+    }],
     id: 'test-id',
     label: 'test label',
   };
@@ -16,18 +25,88 @@ describe('<CustomFieldInputMultipleChoice>', () => {
     expect(renderer.create(<CustomFieldInputMultipleChoice {...requiredProps} />).toJSON()).toMatchSnapshot();
   });
 
+  describe('autocompleter popup', () => {
+    it('opens on click', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('listbox')).toBeInTheDocument();
+    });
+
+    it('closes on escape key', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('listbox')).toBeInTheDocument();
+      fireEvent.keyDown(document.activeElement, { key: 'Escape' });
+      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+    });
+
+    it('does not open when there are no choices', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} choices={[]} />);
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+    });
+
+    it('does not open when read-only', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} readOnly />);
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('listbox')).not.toBeInTheDocument();
+    });
+  });
+
+  describe('deselecting a choice', () => {
+    it('removes the choice value when pressing the remove button', () => {
+      render((<CustomFieldInputMultipleChoice
+        {...requiredProps}
+        value={requiredProps.choices}
+      />));
+
+      expect(screen.getByText('Choice 1')).toBeInTheDocument();
+      expect(screen.getByText('Choice 2')).toBeInTheDocument();
+
+      userEvent.click(screen.getAllByRole('button')[0]);
+      expect(screen.queryByText('Choice 1')).not.toBeInTheDocument();
+      expect(screen.getByText('Choice 2')).toBeInTheDocument();
+    });
+  });
+
+  describe('choices API', () => {
+    it('can be set', () => {
+      render((<CustomFieldInputMultipleChoice
+        {...requiredProps}
+        choices={[{
+          id: '42',
+          label: 'Answer to everything',
+        }]}
+      />));
+
+      expect(screen.queryByRole('option', { name: 'Answer to everything' })).not.toBeInTheDocument();
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.getByRole('option', { name: 'Answer to everything' })).toBeInTheDocument();
+    });
+
+    it('does not show a selected choice', () => {
+      const choices = [{
+        id: '42',
+        label: 'Answer to everything',
+      }];
+      render((<CustomFieldInputMultipleChoice
+        {...requiredProps}
+        choices={choices}
+        value={choices}
+      />));
+
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('option', { name: 'Answer to everything' })).not.toBeInTheDocument();
+    });
+  });
+
   describe('id API', () => {
     it('generates unique tag ids', () => {
       render((<CustomFieldInputMultipleChoice
         {...requiredProps}
         id="123"
-        value={[{
-          id: '1',
-          label: 'Choice 1',
-        }, {
-          id: '2',
-          label: 'Choice 2',
-        }]}
+        value={requiredProps.choices}
       />));
 
       expect(screen.getByText('Choice 1').parentElement).toHaveAttribute('id', '123-1');
@@ -48,10 +127,7 @@ describe('<CustomFieldInputMultipleChoice>', () => {
       render((<CustomFieldInputMultipleChoice
         {...requiredProps}
         readOnly={true}
-        value={[{
-          id: '1',
-          label: 'Choice 1',
-        }]}
+        value={[requiredProps.choices[0]]}
       />));
 
       expect(screen.queryByRole('button')).toBeNull();
@@ -61,13 +137,31 @@ describe('<CustomFieldInputMultipleChoice>', () => {
       render((<CustomFieldInputMultipleChoice
         {...requiredProps}
         readOnly={false}
-        value={[{
-          id: '1',
-          label: 'Choice 1',
-        }]}
+        value={[requiredProps.choices[0]]}
       />));
 
       expect(screen.getByRole('button')).toBeInTheDocument();
+    });
+  });
+
+  describe('selecting a choice', () => {
+    it('removes the choice from the list of choices', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('option', { name: 'Choice 1' })).toBeInTheDocument();
+      userEvent.click(screen.getByRole('option', { name: 'Choice 1' }));
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('option', { name: 'Choice 1' })).not.toBeInTheDocument();
+    });
+
+    it('adds the choice to the selected choices', () => {
+      render(<CustomFieldInputMultipleChoice {...requiredProps} />);
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('gridcell', { name: 'Choice 1' })).not.toBeInTheDocument();
+      userEvent.click(screen.queryByRole('option', { name: 'Choice 1' }));
+      expect(screen.queryByRole('option', { name: 'Choice 1' })).not.toBeInTheDocument();
+      userEvent.click(screen.getByLabelText('test label'));
+      expect(screen.queryByRole('gridcell', { name: 'Choice 1' })).toBeInTheDocument();
     });
   });
 
@@ -75,13 +169,7 @@ describe('<CustomFieldInputMultipleChoice>', () => {
     it('generates tags', () => {
       render((<CustomFieldInputMultipleChoice
         {...requiredProps}
-        value={[{
-          id: '1',
-          label: 'Choice 1',
-        }, {
-          id: '2',
-          label: 'Choice 2',
-        }]}
+        value={requiredProps.choices}
       />));
 
       expect(screen.getByText('Choice 1')).toBeInTheDocument();