+++ src/components/listbox/listbox.jsx
@@ -16,28 +16,6 @@ const Listbox = forwardRef(function Listbox(props, forwardedRef) {
   const backupRef = useRef();
   const ref = forwardedRef || backupRef;
 
-  useEffect(() => {
-    if (active) {
-      props.refs.forEach((optionRef, index) => {
-        if (optionRef.current) optionRef.current.setActive(index === activeIndex);
-      });
-    }
-  });
-
-  useEffect(() => {
-    if (!didMount) return;
-
-    props.onChange({ target: ref.current });
-  }, [value]);
-
-  useEffect(() => {
-    setDidMount(true);
-  }, []);
-
-  useImperativeHandle(ref, () => ({
-    value,
-  }));
-
   const refIndexOf = target => props.refs.findIndex(optionRef => (
     optionRef.current && optionRef.current.contains(target)
   ));
@@ -95,6 +73,29 @@ const Listbox = forwardRef(function Listbox(props, forwardedRef) {
       default:
     }
   }
+
+  useEffect(() => {
+    if (active) {
+      props.refs.forEach((optionRef, index) => {
+        if (optionRef.current) optionRef.current.setActive(index === activeIndex);
+      });
+    }
+  }, [active, activeIndex]);
+
+  useEffect(() => {
+    if (!didMount) return;
+
+    props.onChange({ target: ref.current });
+  }, [value]);
+
+  useEffect(() => {
+    setDidMount(true);
+  }, []);
+
+  useImperativeHandle(ref, () => ({
+    focus: () => props.refs[activeIndex].current.setActive(true),
+    value,
+  }));
 
   return (
     <ul